Parallelized normalised predictive density over vector z

Parallelized normalised predictive density over vector z

ZIB_convolution(
  z_values,
  alpha_input,
  beta_input,
  zi_input,
  weighted_samps,
  weights,
  point
)

Arguments

z_values

evaluation points

alpha_input

Point estimates (point=TRUE) or matrix (point=FALSE) of shape parameters for each element (b) in the aggregate over each of the N observations.

beta_input

Point estimates (point=TRUE) or matrix (point=FALSE) of shape parameters for each element in the aggregate.

zi_input

Point estimates (point=TRUE) or matrix (point=FALSE) of zero-inflation probability parameters for each element in the aggregate.

weighted_samps

Array of weighted samples for each element in the aggregate (dimensions: n_sims x n_draws x b).

weights

vector of weights that are used to combine to form the aggregated density Z. (length b)

point

A true/false indicator to denote whether you are using point estimates (point=TRUE) or posterior samples (point=FALSE) of the Beta parameters.

Value

The aggregate density Z over a grid of values using a convolution of zero-inflated Beta distributions.

Details

A wrapper function for calculating Monte Carlo estimates for the aggregate density Z using the zero-inflated Beta distribution.